requirements.txt
Django==2.0.2
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
jsonfield==2.0.2
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
